style: restructure modules
