+++ flt/pipeline_editor/lib/widgets/pipeline_layer.dart
@@ -1,11 +1,10 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:graph/models/graph_model.dart';
+import 'package:graph/models/pipeline.dart';
+import 'package:graph/models/well.dart';
 import 'package:graph/widgets/pipeline_painter.dart';
 import 'package:latlong2/latlong.dart';
-
 
 class PipelineLayer extends StatefulWidget {
   const PipelineLayer({super.key});
@@ -13,7 +12,6 @@ class PipelineLayer extends StatefulWidget {
   @override
   State<PipelineLayer> createState() => _PipelineLayerState();
 }
-
 
 class _PipelineLayerState extends State<PipelineLayer> {
   int selectionIndex = -1;
@@ -33,7 +31,7 @@ class _PipelineLayerState extends State<PipelineLayer> {
   ];
 
   void onTapDown(TapDownDetails details, MapCamera camera) {
-    for (var (i, n) in  shapes.indexed) {
+    for (var (i, n) in shapes.indexed) {
       if (n.contains(camera, details.globalPosition)) {
         setState(() {
           selectionIndex = i;
@@ -50,7 +48,8 @@ class _PipelineLayerState extends State<PipelineLayer> {
     if (selectionIndex == -1) {
       return;
     }
-    shapes[selectionIndex].latLng.first = camera.pointToLatLng(details.globalPosition.toPoint());
+    shapes[selectionIndex].latLng.first =
+        camera.pointToLatLng(details.globalPosition.toPoint());
     setState(() {
       forceRepaint = !forceRepaint;
     });
@@ -63,23 +62,22 @@ class _PipelineLayerState extends State<PipelineLayer> {
     setState(() {
       forceRepaint = !forceRepaint;
     });
-  }  
+  }
 
   @override
-  Widget build(BuildContext context) { 
+  Widget build(BuildContext context) {
     var camera = MapCamera.of(context);
     var controller = MapController.of(context);
     var options = MapOptions.of(context);
 
     return MobileLayerTransformer(
-      child: GestureDetector(
-        onTapDown: (details) => onTapDown(details, camera),
-        onPanUpdate: editMode ? (details) => onPanUpdate(details, camera) : null,
-        onPanEnd: editMode ? (details) => onPanEnd(details) : null,
-        child: CustomPaint(
-          painter: PipelinePainter(camera, shapes, selectionIndex, forceRepaint),
-        ),
-      )
-    );
+        child: GestureDetector(
+      onTapDown: (details) => onTapDown(details, camera),
+      onPanUpdate: editMode ? (details) => onPanUpdate(details, camera) : null,
+      onPanEnd: editMode ? (details) => onPanEnd(details) : null,
+      child: CustomPaint(
+        painter: PipelinePainter(camera, shapes, selectionIndex, forceRepaint),
+      ),
+    ));
   }
 }
